requirements.txt
streamlit==1.32.0
gspread==6.0.0
google-auth==2.28.2